Refactor: push down subclass-specific methods to subclass
Since there're only two subclasses of StubbedMethod, I didn't see a reason to
privilege methods of one subclass by 'defaulting' them in the super/base class.
So, StubbedMethod is now an abstract class.
