+++ lib/mocha/stubbed_method.rb
@@ -126,26 +126,5 @@ module Mocha
     def stub_method_owner
       @stub_method_owner ||= original_method_owner
     end
-
-    def mock_owner
-      stubbee
-    end
-
-    def original_method_body
-      if PRE_RUBY_V19
-        original_method_in_scope = original_method
-        proc { |*args, &block| original_method_in_scope.call(*args, &block) }
-      else
-        original_method
-      end
-    end
-
-    def store_original_method
-      @original_method = stubbee._method(method_name)
-    end
-
-    def original_method_owner
-      stubbee.singleton_class
-    end
   end
 end
